Added Secret and removed time scale freezing once you win
Added in a secret platform to the top left which has a 5 point pickup that you need to jump the wall for. Also removed the time freezing after you win so that you could look around even if you picked up everything. Also made the ball reset if it fell too far off the map.